+++ govtool/metadata-validation/src/dto/validateMetadata.dto.ts
@@ -1,5 +1,9 @@
+import { MetadataStandard } from '@/types';
+
 export class ValidateMetadataDTO {
   hash: string;
 
   url: string;
+
+  standard?: MetadataStandard;
 }